Run the database-backed tests at last, fix the four bugs they found, and add Ollama as a provider
The 14 database-backed tests had never executed — Docker needs WSL2, WSL needs a reboot — so they were skipped
in every run and quietly rotting. scripts/local-db-native.ps1 (pnpm db:native) removes that dependency: it
fetches the official PostgreSQL binaries (a zip, no installer, no admin), runs a throwaway cluster, builds the
schema from the models and runs pytest. Docker plus `alembic upgrade head` stays the real path — the native one
does not exercise the migration chain and skips `embeddings`, since pgvector is not in a stock PostgreSQL build
and no database-backed test touches it. Both limits are stated in the script rather than left to be discovered.

Running them found four things:

1. tzdata is missing, and this is a production bug rather than a Windows quirk. `zoneinfo` reads the IANA
   database from the OS, and slim and distroless containers ship none — so ZoneInfo("Asia/Kolkata") raises and
   takes the daily check-in rollover and the streak reminder cron with it. tzdata is now a dependency of the API
   and the worker, and settings validation resolves the timezone at boot: falling back to UTC would be worse
   than crashing, because it silently shifts every reward day by five and a half hours for the whole audience.

2. The age-gate test asserted `Conflict` where the API raises `AgeGateRequired` (403 age_gate_required), which
   is what both clients branch on. It could never have passed; nobody saw it because it never ran.

3. Two fixtures left `content_rating` unset, which now means adult, so tests named for ad verification and
   sequential unlocking were really testing the age gate. Ratings are set explicitly, and the seed sets one too
   — otherwise a freshly seeded catalogue is gated for every guest, which is the fail-closed rule biting the
   first person to run it.

4. `series.episodes` was read through the lazy relationship inside async tests, raising MissingGreenlet. The
   fixture now returns its episodes.

The ad-unlock test was also asserting the wrong invariant: it replayed the same episode, which correctly returns
the existing unlock before the ad event is ever consulted. It now pins both behaviours that matter — a replay is
idempotent, and one verified ad completion cannot unlock a second episode, which is the actual fraud case.

61 passed, 0 skipped.

Separately: `ollama` is now a provider for both chat and embeddings, for running the self-hosted server. It is
the one provider that needs no API key — only a reachable base URL — so it suits the high-volume, low-stakes
work (subtitle cleanup, bulk translation drafts) without paying per token. Point individual tasks at it through
KATHA_AI_TASK_PROVIDERS rather than switching the default. mxbai-embed-large is natively 1024-d, so it matches
the pgvector column without padding; switching embedding provider changes model_name() and invalidates existing
vectors, which the env template now says out loud.

+++ services/ai/katha_ai/graphs/subtitles.py
@@ -24,7 +24,7 @@ class Cue:
 
 
 def _ts(seconds: float) -> str:
-    ms = int(round(seconds * 1000))
+    ms = round(seconds * 1000)
     h, rem = divmod(ms, 3_600_000)
     m, rem = divmod(rem, 60_000)
     s, ms = divmod(rem, 1000)
